+++ packages/camera/camera/example/lib/main.dart
@@ -10,6 +10,9 @@ import 'dart:io';
 import 'package:camera/camera.dart';
 import 'package:camera_platform_interface/camera_platform_interface.dart';
 import 'package:camera_platform_interface/method_channel_camera.dart';
+import 'package:camera_platform_interface/camera_controller.dart';
+import 'package:camera_platform_interface/widgets/camera_preview.dart';
+import 'package:camera_platform_interface/exceptions/camera_exception.dart';
 import 'package:flutter/material.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:video_player/video_player.dart';
