+++ packages/camera/camera_platform_interface/lib/exceptions/camera_exception.dart
@@ -1,0 +1,10 @@
+/// This is thrown when the plugin reports an error.
+class CameraException implements Exception {
+  CameraException(this.code, this.description);
+
+  String code;
+  String description;
+
+  @override
+  String toString() => '$runtimeType($code, $description)';
+}
