+++ packages/camera/camera_platform_interface/lib/models/camera_value.dart
@@ -1,0 +1,85 @@
+import 'dart:ui';
+
+/// The state of a [CameraController].
+class CameraValue {
+  const CameraValue({
+    this.isInitialized,
+    this.errorDescription,
+    this.previewSize,
+    this.isRecordingVideo,
+    this.isTakingPicture,
+    this.isStreamingImages,
+    bool isRecordingPaused,
+  }) : _isRecordingPaused = isRecordingPaused;
+
+  const CameraValue.uninitialized()
+      : this(
+    isInitialized: false,
+    isRecordingVideo: false,
+    isTakingPicture: false,
+    isStreamingImages: false,
+    isRecordingPaused: false,
+  );
+
+  /// True after [CameraController.initialize] has completed successfully.
+  final bool isInitialized;
+
+  /// True when a picture capture request has been sent but as not yet returned.
+  final bool isTakingPicture;
+
+  /// True when the camera is recording (not the same as previewing).
+  final bool isRecordingVideo;
+
+  /// True when images from the camera are being streamed.
+  final bool isStreamingImages;
+
+  final bool _isRecordingPaused;
+
+  /// True when camera [isRecordingVideo] and recording is paused.
+  bool get isRecordingPaused => isRecordingVideo && _isRecordingPaused;
+
+  final String errorDescription;
+
+  /// The size of the preview in pixels.
+  ///
+  /// Is `null` until  [isInitialized] is `true`.
+  final Size previewSize;
+
+  /// Convenience getter for `previewSize.height / previewSize.width`.
+  ///
+  /// Can only be called when [initialize] is done.
+  double get aspectRatio => previewSize.height / previewSize.width;
+
+  bool get hasError => errorDescription != null;
+
+  CameraValue copyWith({
+    bool isInitialized,
+    bool isRecordingVideo,
+    bool isTakingPicture,
+    bool isStreamingImages,
+    String errorDescription,
+    Size previewSize,
+    bool isRecordingPaused,
+  }) {
+    return CameraValue(
+      isInitialized: isInitialized ?? this.isInitialized,
+      errorDescription: errorDescription,
+      previewSize: previewSize ?? this.previewSize,
+      isRecordingVideo: isRecordingVideo ?? this.isRecordingVideo,
+      isTakingPicture: isTakingPicture ?? this.isTakingPicture,
+      isStreamingImages: isStreamingImages ?? this.isStreamingImages,
+      isRecordingPaused: isRecordingPaused ?? _isRecordingPaused,
+    );
+  }
+
+  @override
+  String toString() {
+    return '$runtimeType('
+        'isRecordingVideo: $isRecordingVideo, '
+        'isRecordingVideo: $isRecordingVideo, '
+        'isInitialized: $isInitialized, '
+        'errorDescription: $errorDescription, '
+        'previewSize: $previewSize, '
+        'isStreamingImages: $isStreamingImages)';
+  }
+}
